scanner/requirements.txt
python3-nmap==1.9.1
PyYAML==6.0.2
simplejson==3.20.1
